requirements.txt
asgiref==3.9.1
Django==5.2.5
django-ranged-response==0.2.0
django-simple-captcha==0.6.2
Markdown==3.8.2
pillow==11.3.0
sqlparse==0.5.3
typing_extensions==4.14.1
